Add option for expand/collapse directory if disabled.
By default, this functionality is left on to preserve existing
functionality.

I would prefer if I had the option to not change the default
functionality such that clicking a node does not change from a
select/deselect action to expand/collapse if that node is a directory
that is disabled.